requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
sqlalchemy>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
plyer>=2.1.0
reportlab>=4.0.0
pyinstaller>=6.0.0
anime-js-python>=1.0.0
lottie>=0.6.11
popmotion>=9.0.0 